feat : 4단계 기능 구현
