+++ src/main/java/racinggame/Main.java
@@ -1,0 +1,9 @@
+package racinggame;
+
+import racinggame.controller.RacingController;
+
+public class Main {
+    public static void main(String[] args) {
+        new RacingController().run();
+    }
+}
