+++ src/main/java/racinggame/policy/decider/MoveDecider.java
@@ -1,4 +1,4 @@
-package car.move.decider;
+package racinggame.policy.decider;
 
 public interface MoveDecider {
     boolean canMove(int randomNumber);
