+++ src/main/java/racinggame/policy/strategy/MoveStrategy.java
@@ -1,4 +1,4 @@
-package car.move.strategy;
+package racinggame.policy.strategy;
 
 public interface MoveStrategy {
     int addStepSize();
